Split monolithic gear test file into per-component test files; add testUtils/renderUtils.tsx
Agent-Logs-Url: https://github.com/CptnFizzbin/shadow-sin/sessions/44f9d1d4-b29e-489b-9612-184684127cd3
