+++ src/components/characterBuilder/sections/gear/cyberware/dialogs/implantFormDialog.test.tsx
@@ -1,0 +1,22 @@
+import { waitFor } from "@testing-library/react"
+import { describe, expect, it, vi } from "vitest"
+
+import { ImplantFormDialog } from "#/components/characterBuilder/sections/gear/cyberware/dialogs/implantFormDialog.tsx"
+import type { ImplantData } from "#/lib/system/gear/implantData.ts"
+import { ItemType } from "#/lib/system/itemType.ts"
+import { fillNameAndClickSave, renderWithTheme } from "#testUtils/renderUtils.tsx"
+
+describe("ImplantFormDialog", () => {
+  it("submits an item with ItemType.implant", async () => {
+    const onSave = vi.fn()
+    renderWithTheme(<ImplantFormDialog open onSave={onSave} onClose={vi.fn()} />)
+
+    fillNameAndClickSave("Wired Reflexes 1")
+
+    await waitFor(() => {
+      expect(onSave).toHaveBeenCalledOnce()
+      const submitted: ImplantData = onSave.mock.calls[0][0]
+      expect(submitted.itemType).toBe(ItemType.implant)
+    })
+  })
+})
