+++ src/components/characterBuilder/sections/gear/generic/forms/useItemForm.test.tsx
@@ -1,0 +1,40 @@
+import { renderHook } from "@testing-library/react"
+import { describe, expect, it, vi } from "vitest"
+
+import { useItemForm } from "#/components/characterBuilder/sections/gear/generic/forms/useItemForm.tsx"
+import type { ItemData } from "#/lib/system/itemData.ts"
+import { ItemType } from "#/lib/system/itemType.ts"
+
+describe("useItemForm", () => {
+  it("defaults to ItemType.other when no itemType is provided", () => {
+    const { result } = renderHook(() => useItemForm({ onSubmit: vi.fn() }))
+    expect(result.current.state.values.itemType).toBe(ItemType.other)
+  })
+
+  it("uses ItemType.vehicle when itemType=vehicle is provided", () => {
+    const { result } = renderHook(() =>
+      useItemForm({ itemType: ItemType.vehicle, onSubmit: vi.fn() }),
+    )
+    expect(result.current.state.values.itemType).toBe(ItemType.vehicle)
+  })
+
+  it("uses ItemType.armor when itemType=armor is provided", () => {
+    const { result } = renderHook(() =>
+      useItemForm({ itemType: ItemType.armor, onSubmit: vi.fn() }),
+    )
+    expect(result.current.state.values.itemType).toBe(ItemType.armor)
+  })
+
+  it("preserves the existing item's itemType when editing", () => {
+    const existingItem: ItemData = {
+      id: crypto.randomUUID(),
+      itemType: ItemType.vehicle,
+      name: "Eurocar Westwind 2000",
+      cost: 65000,
+    }
+    const { result } = renderHook(() =>
+      useItemForm({ item: existingItem, onSubmit: vi.fn() }),
+    )
+    expect(result.current.state.values.itemType).toBe(ItemType.vehicle)
+  })
+})
