+++ src/components/characterBuilder/sections/gear/weapons/forms/useWeaponForm.test.tsx
@@ -1,0 +1,12 @@
+import { renderHook } from "@testing-library/react"
+import { describe, expect, it, vi } from "vitest"
+
+import { useWeaponForm } from "#/components/characterBuilder/sections/gear/weapons/forms/useWeaponForm.tsx"
+import { ItemType } from "#/lib/system/itemType.ts"
+
+describe("useWeaponForm", () => {
+  it("always defaults to ItemType.weapon", () => {
+    const { result } = renderHook(() => useWeaponForm({ onSubmit: vi.fn() }))
+    expect(result.current.state.values.itemType).toBe(ItemType.weapon)
+  })
+})
